backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic-settings>=2,<3
supabase>=2,<3
dnspython>=2.7,<3
sendgrid>=6,<7
apscheduler>=3.10,<4
python-dotenv>=1,<2
